# WelcomeHome/WelcomeHome.Web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WelcomeHome.Web",
    "version": "1.0"
  },
  "paths": {
...
    "/Step/{id}": {
      "get": {
        "tags": [
          "Step"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StepOutDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepOutDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepOutDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Step"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Step": {
      "get": {
        "tags": [
          "Step"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Step"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepInDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StepInDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StepInDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Step"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepOutDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StepOutDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StepOutDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/byEstablishmentType/{establishmentTypeId}": {
      "get": {
        "tags": [
          "Step"
        ],
        "parameters": [
          {
            "name": "typeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "establishmentTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepOutDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "City": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          }
        },
        "additionalProperties": false
      },
...
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/City"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DocumentOutDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EstablishmentOutDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "pageURL": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "otherContacts": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "$ref": "#/components/schemas/City"
          },
          "establishmentType": {
            "$ref": "#/components/schemas/EstablishmentTypeOutDTO"
          }
        },
        "additionalProperties": false
      },
      "EstablishmentTypeOutDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StepInDTO": {
        "type": "object",
        "properties": {
          "sequenceNumber": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "establishmentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "documentsBringId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "documentsReceiveId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StepOutDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "sequenceNumber": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "establishments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstablishmentOutDTO"
            },
            "nullable": true
          },
          "documentsBring": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentOutDTO"
            },
            "nullable": true
          },
          "documentsReceive": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentOutDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}